Template answer view (TODO: fix hard code)
